# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: MyASA 2.0 API
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /activities:
    get:
      operationId: getActivities
      tags: [activities]
      summary: List recurring/one-off activities for manageable operations
      parameters:
        - name: operationId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Activities list
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      $ref: "#/components/schemas/Activity"
                required: [activities]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create a recurring or one-off activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityCreate"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                type: object
                properties:
                  activity:
                    $ref: "#/components/schemas/Activity"
                required: [activity]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity and (optionally) its assignees
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Activity updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  activity:
                    $ref: "#/components/schemas/Activity"
                required: [activity]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activity deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                required: [ok]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    ActivitySchedule:
      type: object
      required: [id]
      properties:
        id: { type: string }
        weekday: { type: integer, nullable: true }
        specificDate: { type: string, format: date, nullable: true }
        startTime: { type: string, nullable: true }
        endTime: { type: string, nullable: true }
    ActivityScheduleInput:
      type: object
      properties:
        weekday: { type: integer, nullable: true }
        specificDate: { type: string, format: date, nullable: true }
        startTime: { type: string, nullable: true }
        endTime: { type: string, nullable: true }
    ActivityAssignee:
      type: object
      required: [id]
      properties:
        id: { type: string }
        userId: { type: string, nullable: true }
        groupId: { type: string, nullable: true }
        userName: { type: string, nullable: true }
        groupName: { type: string, nullable: true }
    Activity:
      type: object
      required: [id, organizationId, operationId, title, active, schedules, assignees, createdAt, updatedAt]
      properties:
        id: { type: string }
        organizationId: { type: string }
        operationId: { type: string }
        title: { type: string }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        schedules:
          type: array
          items:
            $ref: "#/components/schemas/ActivitySchedule"
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/ActivityAssignee"
    ActivityAssigneeInput:
      type: object
      properties:
        userId: { type: string, nullable: true }
        groupId: { type: string, nullable: true }
    ActivityCreate:
      type: object
      required: [operationId, title, schedules]
      properties:
        operationId: { type: string }
        title: { type: string }
        active: { type: boolean }
        schedules:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ActivityScheduleInput"
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/ActivityAssigneeInput"
    ActivityUpdate:
      type: object
      properties:
        title: { type: string }
        active: { type: boolean }
        schedules:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ActivityScheduleInput"
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/ActivityAssigneeInput"
...